Include flow types into library (#649)
* Fix types for modern flow

* Move flow declarations and test into recompose

* Fix docs

* Remove errors

* Add yarn-error to gitignore

* Append flow 2 test

* Update release to support flow types

* Remove unneeded command
