+++ .gitignore
@@ -3,3 +3,4 @@ release
 lib
 coverage
 .vscode
+yarn-error.log
