+++ src/packages/recompose/__tests__/types/test_classBasedEnhancer.js
@@ -1,7 +1,7 @@
 /* @flow */
 import * as React from 'react'
-import { compose, withProps } from 'recompose'
-import type { HOC } from 'recompose'
+import { compose, withProps } from '../..'
+import type { HOC } from '../..'
 
 // Example of very dirty written fetcher enhancer
 function fetcher<Response: {}, Base: {}>(
