+++ src/packages/recompose/__tests__/types/test_onlyUpdateForPropTypes.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions */
 /* @flow */
 import React from 'react'
-import { compose, withProps, pure } from 'recompose'
+import { compose, withProps, onlyUpdateForPropTypes } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
 
@@ -17,7 +17,7 @@ const Comp = ({ eA }) =>
   </div>
 
 const enhacer: HOC<*, EnhancedCompProps> = compose(
-  pure,
+  onlyUpdateForPropTypes,
   withProps(props => ({
     eA: (props.eA: number),
     // $ExpectError eA nor any nor string
