+++ src/packages/recompose/__tests__/types/test_voodoo.js
@@ -9,9 +9,9 @@ import {
   renameProp,
   renameProps,
   withState,
-} from 'recompose'
+} from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = {
   eA: number,
