+++ src/packages/recompose/__tests__/types/test_withProps.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions */
 /* @flow */
 import React from 'react'
-import { compose, withProps } from 'recompose'
+import { compose, withProps } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { a: string, b: number }
