+++ src/packages/recompose/__tests__/types/test_withStateHandlers.js
@@ -2,9 +2,9 @@
 /* @flow */
 
 import React from 'react'
-import { compose, withProps, withStateHandlers } from 'recompose'
+import { compose, withProps, withStateHandlers } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = {
   initialCounter: number,
@@ -12,7 +12,7 @@ type EnhancedCompProps = {
 
 const enhancer: HOC<*, EnhancedCompProps> = compose(
   withStateHandlers(
-    { value: 'world', letIt: 'be', obj: ({}: { [key: string]: string }) },
+    { value: 'Hello', letIt: 'be', obj: ({}: { [key: string]: string }) },
     {
       // we need to set argument type so inference will work good
       setValue: (state, props) => (value: string) => ({
@@ -66,12 +66,13 @@ const enhancerFuncInit: HOC<*, EnhancedCompProps> = compose(
   }))
 )
 
-const BaseComponent = ({ hi, changeValue }) =>
+const BaseComponent = ({ hi, changeValue, setValue }) =>
   <div
     onClick={() => {
       // check that supports few arguments
       const x = changeValue({ i: 1, j: '1' }, 1)
 
+      setValue('ww')
       // Check that result is void
       ;(x: void)
