--- types/flow-typed/README.md
@@ -1,13 +1,0 @@
-# flow-typed definitions for recompose
-
-Up to date definitions for recompose package.
-Must be synchronized to [flow-typed](https://github.com/flowtype/flow-typed)
-
-## Contribute
-
-```shell
-# git clone recompose
-cd recompose/examples/flow-typed
-yarn
-yarn flow
-```
