--- types/flow-typed/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "recompose-flow-typed",
-  "version": "1.0.0",
-  "main": "index.js",
-  "author": "istarkov@gmail.com",
-  "license": "MIT",
-  "devDependencies": {
-    "flow-bin": "^0.55.0",
-    "prop-types": "^15.5.10",
-    "rxjs": "^5.4.2"
-  },
-  "scripts": {
-    "flow": "flow"
-  }
-}
